test: 삭제된 댓글을 다시 삭제하려는 예외 추가 및 테스트 코드 작성

+++ week4/src/main/java/com/kyle/week4/service/CommentService.java
@@ -73,6 +73,10 @@ public class CommentService {
             throw new CustomException(PERMISSION_DENIED);
         }
 
+        if (comment.isDeleted()) {
+            throw new CustomException(ALREADY_DELETED_COMMENT);
+        }
+
         comment.delete();
         commentCountRepository.decrease(postId);
     }
